reorganize CorlaViewer add SamplingPanel rejigger CorlaContestsTable, CountyTable delete CountyPoolsTable - functionality transfered to other tables
