+++ viewer/src/main/kotlin/org/cryptobiotic/rlauxe/viewer/CountyTable.kt
@@ -1,33 +1,38 @@
 /*
- * Copyright (c) 2025 John L. Caron
+ * Copyright (c) 2026 John L. Caron
  * See LICENSE for license information.
  */
 package org.cryptobiotic.rlauxe.viewer
 
-import org.cryptobiotic.rlauxe.audit.ContestRound
-import org.cryptobiotic.rlauxe.betting.TestH0Status
-import org.cryptobiotic.rlauxe.betting.estRiskStandardBet
-import org.cryptobiotic.rlauxe.betting.estSampleSizeStandardBet
-import org.cryptobiotic.rlauxe.core.ContestWithAssertions
+import org.cryptobiotic.rlauxe.audit.CountyPools
+import org.cryptobiotic.rlauxe.core.ContestInfo
+import org.cryptobiotic.rlauxe.estimate.Vunder
 import org.cryptobiotic.rlauxe.persist.AuditRecord.Companion.read
 import org.cryptobiotic.rlauxe.persist.CountyAuditRecord
-import org.cryptobiotic.rlauxe.persist.CountyContestData
 import org.cryptobiotic.rlauxe.persist.CountyData
+import org.cryptobiotic.rlauxe.util.ContestTabulation
 import org.cryptobiotic.rlauxe.util.dfn
 import org.cryptobiotic.rlauxe.util.nfn
+import org.cryptobiotic.rlauxe.util.trunc
+import org.cryptobiotic.rlauxe.workflow.PersistedMvrManager
 import org.slf4j.Logger
 import org.slf4j.LoggerFactory
 import ucar.ui.prefs.BeanTable
+import ucar.ui.prefs.BeanTable.TableBeanProperty
+import ucar.ui.widget.BAMutil
 import ucar.ui.widget.IndependentWindow
 import ucar.ui.widget.TextHistoryPane
 import ucar.util.prefs.PreferencesExt
 import java.awt.BorderLayout
+import java.awt.event.ActionEvent
 import java.util.function.Function
-import javax.swing.JOptionPane
+import javax.swing.AbstractAction
 import javax.swing.JPanel
 import javax.swing.JSplitPane
 import javax.swing.event.ListSelectionEvent
 import javax.swing.event.ListSelectionListener
+
+private val logger: Logger = LoggerFactory.getLogger(ContestPoolsTable::class.java)
 
 class CountyTable(
     private val prefs: PreferencesExt,
@@ -35,307 +40,323 @@ class CountyTable(
     infoWindow: IndependentWindow?,
     fontSize: Float,
 ) : JPanel(), ViewerPanelIF {
-    private val countyTable: BeanTable<CountyBean>
-    private val contestTable: BeanTable<CountyContestBean>
+    private val countyTable: BeanTable<CountyPoolsBean>
+    private val styleTable: BeanTable<StyleTable.StyleBean>
+    private val contestTable: BeanTable<CountyPoolContestBean>
 
+    // TextHistoryPane localInfo = new TextHistoryPane();
+    private val split1: JSplitPane
     private val split2: JSplitPane
 
     private var auditRecordLocation: String? = "none"
-    private var countyAudit: CountyAuditRecord? = null
-    // private var config: Config? = null
-
-    var countyMap = emptyMap<String, CountyBean>()
-    var totalBean : CountyBean? = null
-
-    var contestMap = emptyMap<Int, ContestWithAssertions>()
-    var contestRoundMap = emptyMap<Int, ContestRound>()
-    var countyContestData = emptyList<CountyContestData>()
-    var auditRiskLimit = .03
+    private var countyRecord: CountyAuditRecord? = null
+    var mvrManager: PersistedMvrManager? = null
+    var infos: Map<Int, ContestInfo> = emptyMap()
+    var countyCvrMap: Map<String, CountyPools> = emptyMap()
 
     init {
         countyTable = BeanTable(
-            CountyBean::class.java, prefs.node("countyTable") as PreferencesExt?, false,
-            "Counties", "County Audit", null
+            CountyPoolsBean::class.java, prefs.node("countyCardPool") as PreferencesExt?, false,
+            "CountyPools", "CountyPools", null
+        )
+        countyTable.addPopupOption(
+            "Show CountyPool",
+            countyTable.makeShowAction(infoTA, infoWindow, Function { bean: Any? -> showCountyPool(bean as CountyPoolsBean) })
+        )
+        countyTable.addPopupOption(
+            "Print Table", countyTable.makeShowAction(infoTA, infoWindow,
+                Function { bean: Any? ->
+                    val indexBeans : List<BeanTable.IndexedBean<CountyPoolsBean>> = countyTable.getIndexedBeans()
+                    val sortedIndexBeans : List<BeanTable.IndexedBean<CountyPoolsBean>> = indexBeans.sortedBy { it.viewIndex() }
+                    val sortedBeans : List<CountyPoolsBean> =   sortedIndexBeans.map { it.bean }
+                    BeanProperties.printTableG(sortedBeans, countyTable.tableModel, CountyPoolsBean.beanProperties,"countyPool")
+                }
+            )
         )
         countyTable.addListSelectionListener(ListSelectionListener { e: ListSelectionEvent? ->
-            val county = countyTable.getSelectedBean()
-            if (county != null) {
-                setSelectedCounty(county)
+            val selected = countyTable.getSelectedBean()
+            if (selected != null) {
+                setSelectedCounty(selected)
             }
         })
 
         contestTable = BeanTable(
-            CountyContestBean::class.java, prefs.node("countyContestTable") as PreferencesExt?, false,
-            "CountyContest", "CountyContest", null
+            CountyPoolContestBean::class.java, prefs.node("contestTable") as PreferencesExt?, false,
+            "Contest subtotal in selected County", "ContestTabulation", null
         )
         contestTable.addPopupOption(
-            "Show Contest", contestTable.makeShowAction(
-                infoTA, infoWindow,
-                Function { bean: Any? -> showContest((bean as CountyContestBean)) })
+            "Show County Contest", contestTable.makeShowAction(infoTA, infoWindow,
+                Function { bean: Any? -> showCountyContest(bean as CountyPoolContestBean) })
+        )
+
+        styleTable = BeanTable(
+            StyleTable.StyleBean::class.java, prefs.node("styleTable") as PreferencesExt?, false,
+            "Style", "Style", null
+        )
+        styleTable.addPopupOption(
+            "Show use in Contests", styleTable.makeShowAction(infoTA, infoWindow,
+                Function { bean: Any? -> showCountyStyle(bean as StyleTable.StyleBean) })
         )
 
         setFontSize(fontSize)
 
         // layout of tables
-        split2 = JSplitPane(JSplitPane.VERTICAL_SPLIT, false, countyTable, contestTable)
-        split2.setDividerLocation(prefs.getInt("splitPos2", 200))
+        split1 = JSplitPane(JSplitPane.VERTICAL_SPLIT, false, countyTable, contestTable)
+        split1.setDividerLocation(prefs.getInt("splitPos1", 200))
+        split2 = JSplitPane(JSplitPane.VERTICAL_SPLIT, false, split1, styleTable)
+        split2.setDividerLocation(prefs.getInt("splitPos2", 600))
 
-        // auditRoundTable not used for now
         setLayout(BorderLayout())
         add(split2, BorderLayout.CENTER)
 
-        logger.debug("corlaPanel init")
+        logger.debug("CountyPoolTable init")
     }
 
-    fun setSelectedTab() {
-        countMvrsByCounty() // TODO does this mean we dont need the button ??
-    }
-
-    fun countMvrsByCounty() {
-        var countMvrs = 0
-        val mvrCounts = countyAudit!!.countMvrsByCounty() // mvr counts (cardStyle sampling)
-        for (countyData in mvrCounts.values) {
-            val countyBean = countyMap.get(countyData.countyName)
-            if (countyBean != null) {
-                countyBean.nmvrsCardStyleSampling = countyData.nmvrs
-                countMvrs += countyData.nmvrs
-            } else {
-                logger.warn("cant find countyName '" + countyData.countyName + "'")
+    // actions on right side of Audit record chooser
+    fun getActions(container: JPanel) {
+        val readMvrAction: AbstractAction = object : AbstractAction() {
+            override fun actionPerformed(e: ActionEvent?) {
+                val selected = countyTable.getSelectedBean()
+                if (selected != null) {
+                    // readMvrTabulation(selected)
+                }
             }
         }
-        if (totalBean != null) {
-            totalBean!!.nmvrsCardStyleSampling = countMvrs
-        }
-        countyTable.refresh()
+        BAMutil.setActionProperties(readMvrAction, "sunrise-icon.png", "Read Mvr Tabulation", false, 'T'.code, -1)
+        BAMutil.addActionToContainer(container, readMvrAction)
     }
 
     override fun setFontSize(size: Float) {
-        contestTable.setFontSize(size)
         countyTable.setFontSize(size)
+        contestTable.setFontSize(size)
+        styleTable.setFontSize(size)
     }
 
     override fun setAuditRecord(auditRecordLocation: String): Boolean {
+        logger.debug("ContestPoolsTable setAuditRecord " + auditRecordLocation)
+        countyTable.setBeans(emptyList())
+        contestTable.setBeans(emptyList())
+        styleTable.setBeans(emptyList())
+
         this.auditRecordLocation = auditRecordLocation
-        contestTable.setBeans(mutableListOf<CountyContestBean?>())
-
-        logger.debug("countyPanel setAuditRecord " + auditRecordLocation)
-
-        this.auditRecordLocation = auditRecordLocation
-        val record = read(auditRecordLocation)
-        if (record == null) return false
-        if (record !is CountyAuditRecord) return false
-
-        this.countyAudit = record
-        val config = countyAudit!!.config
-        this.auditRiskLimit = config.riskLimit
-        val lastAuditRound = countyAudit!!.rounds.last()
-
-        try {
-            var countUniformMvrs = 0
-            var statewide: CountyBean? = null
-
-            val countyList = mutableListOf<CountyBean>()
-            for (countyData in countyAudit!!.countyData) {
-                val bean = CountyBean(countyData)
-                if (bean.name == "Statewide") statewide = bean
-                    else countUniformMvrs += bean.nmvrsUniformSampling // statewide now included in counties, so dont count twice
-
-                countyList.add(bean)
-            }
-            val totalPopulation = countyAudit!!.countyData.filter { it.countyName != "Statewide"}.sumOf { it.npop }
-            val totalBean = CountyBean( CountyData("=Total", countUniformMvrs, totalPopulation))
-
-            countyList.add(totalBean)
-            countyTable.setBeans(countyList)
-            this.totalBean = totalBean
-
-            // if (statewide != null) statewide.nmvrsUniform = countUniformMvrs;
-            this.countyMap = countyList.associateBy { it.name }
-            this.countyContestData = countyAudit!!.countyContestData
-            countMvrsByCounty()
-
-            this.contestMap =  countyAudit!!.contests.associateBy { it.id }
-            this.contestRoundMap =  lastAuditRound.contestRounds.associateBy { it.id }
-
-        } catch (e: Exception) {
-            e.printStackTrace()
-            JOptionPane.showMessageDialog(null, e.message)
-            logger.error("setAuditRecord failed", e)
+        val auditRecord = read(auditRecordLocation)
+        if (auditRecord == null) {
+            logger.info("ContestPoolsTable failed on readFrom " + auditRecordLocation)
+            return false
         }
+
+        if (auditRecord !is CountyAuditRecord) return false
+        this.countyRecord = auditRecord as CountyAuditRecord
+        this.mvrManager = PersistedMvrManager(this.countyRecord!!, false)
+
+        val countyPools = mvrManager!!.countyPools()
+        if (countyPools == null) return false
+
+        val countyCvrs = mvrManager!!.countyCvrPools() //.associateBy { it.countyName }
+        this.countyCvrMap = if (countyCvrs == null) emptyMap() else countyCvrs.associateBy { it.countyName }
+        logger.info("read countyCvrPools=" + countyCvrMap.size)
+
+        this.infos = auditRecord.contests.associate { it.contest.info().id to it.contest.info() }
+        val countyData = countyRecord!!.countyData.associateBy { it.countyName }
+
+        val beanList = mutableListOf<CountyPoolsBean>()
+        countyPools.forEach {
+            val bean = CountyPoolsBean(it, countyData[it.countyName]!!)
+            val cvrTabs = this.countyCvrMap[ it.countyName ]
+            if (cvrTabs != null)
+                bean.cvrTabs = cvrTabs.contestTabs
+            beanList.add(bean)
+        }
+        countyTable.setBeans(beanList)
 
         return true
     }
 
-    fun setSelectedCounty(countyBean: CountyBean) {
-        val beanList = mutableListOf<CountyContestBean>()
-        for (countyContest in countyContestData) {
-            val contestRound = contestRoundMap[countyContest.id]
-            if (contestRound != null && contestRound.status == TestH0Status.InProgress) {
-                if (countyContest.countyName == countyBean.name) {
-                    val bean = CountyContestBean(countyContest, this.auditRiskLimit)
-                    beanList.add(bean)
-                    bean.countyPopulation = countyBean.population
-                    bean.nmvrsConsistent = countyBean.nmvrsCardStyleSampling
-                    bean.nmvrsUniform = countyBean.nmvrsUniformSampling
+    fun setSelectedCounty(countyBean: CountyPoolsBean) {
+        contestTable.setBeans(emptyList())
+        styleTable.setBeans(emptyList())
 
-                    bean.contestUA = contestMap[countyContest.id]
-                    bean.contestRound = contestRound
-                }
+        val styleNCards = mutableMapOf<Int, Int>()
+
+        val sbeanList = mutableListOf<StyleTable.StyleBean>()
+        countyBean.countyPool.styles.forEach { style ->
+            sbeanList.add( StyleTable.StyleBean(style))
+
+            style.possibleContests().forEach { contestId ->
+                val contestNcard: Int = styleNCards.getOrPut(contestId) { 0 }
+                styleNCards[contestId] = contestNcard + style.ncards()
+            }
+        }
+        styleTable.setBeans(sbeanList)
+
+        val beanList = mutableListOf<CountyPoolContestBean>()
+        countyBean.countyPool.contestTabs.forEach { (id, tab) ->
+            val info = this.infos[id]!!
+            val auditcenterBean = CountyPoolContestBean(countyBean, info, tab, false)
+            beanList.add( auditcenterBean )
+            val cvrTab = countyBean.cvrTabs[id]
+            if (cvrTab != null) {
+                val mvrBean = CountyPoolContestBean(countyBean, info, cvrTab, true)
+                mvrBean.acBean = auditcenterBean
+                beanList.add(mvrBean)
             }
         }
         contestTable.setBeans(beanList)
     }
 
-    fun showContest(contestBean: CountyContestBean) = buildString {
-        append(contestTable.tableModel.showBean(contestBean, BeanProperties.contests))
-        append(contestBean.contestUA!!.show())
-    }
-
     override fun saveState() {
-        contestTable.saveState(false)
         countyTable.saveState(false)
+        styleTable.saveState(false)
+        contestTable.saveState(false)
+
+        prefs.putInt("splitPos1", split1.getDividerLocation())
         prefs.putInt("splitPos2", split2.getDividerLocation())
     }
 
-    companion object {
-        private val logger: Logger = LoggerFactory.getLogger(CountyTable::class.java)
-    }
-}
-
-////////////////////////////////////////////////////////////////
-class CountyBean(countyData: CountyData) {
-    val name: String
-    val population: Int
-    val nmvrsUniformSampling: Int
-    var nmvrsCardStyleSampling: Int = 0
-
-    init {
-        this.name = countyData.countyName
-        this.population = countyData.npop
-        this.nmvrsUniformSampling = countyData.nmvrs
+    fun showCountyContest(bean: CountyPoolContestBean) = buildString {
+        appendLine(contestTable.tableModel.showBean(this, CountyPoolContestBean.beanProperties))
+        appendLine(bean.contestTab)
+        appendLine(bean.vunderTab)
+        appendLine()
+        appendLine("Used in Styles:")
+        styleTable.getBeans().forEach { styleBean ->
+            if (styleBean.style.hasContest(bean.contestId)) {
+                appendLine("  ${styleBean.style}")
+            }
+        }
     }
 
-    fun getDiff() = nmvrsUniformSampling - nmvrsCardStyleSampling
-}
+    fun showCountyStyle(bean: StyleTable.StyleBean) = buildString {
+        appendLine(bean.style)
+        appendLine("Used in Contests:")
+        appendLine(CountyPoolContestBean.header)
 
-// data class CountyContestData(val countyName: String, val contestName: String, val id: Int, val voteDiff: Int, val votes: Map<Int, Int>)
-class CountyContestBean(val countyContestData: CountyContestData, val auditRiskLimit: Double) {
-    var countyPopulation = 1
-    var nmvrsConsistent = 0
-    var nmvrsUniform = 0
-    var contestUA: ContestWithAssertions? = null
-    var contestRound: ContestRound? = null
-
-    // kotlin adds getters, do have to remove the ones that shouldnt be in the table
-
-    fun getCounty() = countyContestData.countyName
-    fun getContestName() = countyContestData.contestName
-    fun getContestId() = countyContestData.id
-    fun getVoteDiff() = countyContestData.voteDiff
-    fun getVotes() = countyContestData.votes.toString()
-
-    fun getContestPopulation() = contestUA?.population() ?: 1
-
-    fun getMargin() = countyContestData.voteDiff / getContestPopulation().toDouble()
-
-    fun getNoerror(): Double {
-        if (contestUA == null) return 0.0
-        val cua = contestUA!!
-        val noerror = cua.minNoerror()
-        return noerror?: 0.0
+        contestTable.getBeans().forEach { contestBean ->
+            if (bean.style.hasContest(contestBean.contestId)) {
+                appendLine(contestBean.show())
+            }
+        }
     }
 
-    fun getNCounties(): String {
-        if (contestUA == null) return "N/A"
-        val CORLAcounties = contestUA!!.contest.info().metadata.get("CORLAcounties")
-        if (CORLAcounties == null) return "N/A"
-        val toks: List<String> = CORLAcounties.split(",".toRegex()).dropLastWhile { it.isEmpty() }
-        if (toks.size == 1) return toks[0]
-        return String.format("%02d", toks.size)
+    fun showCountyPool(bean: CountyPoolsBean) = buildString {
+        appendLine(countyTable.tableModel.showBean(bean, CountyPoolsBean.beanProperties))
+        append(bean.show())
+    }
+    ////////////////////////////////////////////////////////////////
+    // data class CountyPoolMultipleStyles (
+    //    val countyName: String,
+    //    val countyPoolId: Int,
+    //    val contestTabs: List<ContestTabulation>,  // contestId -> ContestTabulation
+    //    val totalCards: Int,
+    //    val cardStyles: List<StyleIF>,
+    //    // val cardStylesCount: List<Int>, // or CardStyleWithNCards ??
+    //)
+    class CountyPoolsBean(val countyPool: CountyPools, val countyData: CountyData) {
+        val countyName = countyPool.countyName
+        val countyPoolId = countyPool.countyPoolId
+        val totalCards = countyPool.cardCount
+
+        // val nmvrs = countyData.nmvrs
+        val population = countyData.npop
+        val diffCards = population - totalCards
+        val diffCardsPct = (population - totalCards) / population.toDouble()
+
+        var cvrTabs: Map<Int, ContestTabulation> = emptyMap()
+
+        val acNvotes = countyPool.contestTabs.values.sumOf { it.nvotes() }
+        fun getCvrNvotes() = cvrTabs.values.sumOf { it.nvotes() }
+
+        fun getDiffNvotes() : Int {
+            return (acNvotes - getCvrNvotes())
+        }
+        fun getPctDiffNvotes() : Double {
+            return (acNvotes - getCvrNvotes()) / acNvotes.toDouble()
+        }
+
+        fun show() = buildString {
+            appendLine(countyData.toString())
+            val totalCards = countyPool.styles.sumOf { it.ncards() }
+            appendLine("sum of countyPool.styles.ncards() = ${totalCards} ")
+        }
+
+        companion object {
+            @JvmStatic
+            fun hiddenProperties() = "countyPool countyData cvrTabs"
+
+            @JvmStatic
+            val beanProperties = listOf(
+                TableBeanProperty("countyName", "county name"),
+                TableBeanProperty("countyPoolId", "county name"),
+                TableBeanProperty("population", "county population from round.ballotCardCount"),
+                TableBeanProperty("totalCards", "number of cvrs in the pool"),
+                TableBeanProperty("diffCards", "population - totalCards"),
+                TableBeanProperty("diffCardsPct", "(population - totalCards)/population"),
+                TableBeanProperty("acNvotes", "auditcenter.votes"),
+                TableBeanProperty("cvrNvotes", "cvr.votes"),
+                TableBeanProperty("diffNvotes", "auditcenter.votes - cvr.votes"),
+                TableBeanProperty("pctDiffNvotes", "(auditcenter.votes - cvr.votes)/auditcenter.votes"),
+            )
+        }
     }
 
-    fun getCorlaSample(): String {
-        if (contestUA == null) return "N/A"
-        val CORLAsample = contestUA!!.contest.info().metadata.get("CORLAsample")
-        if (CORLAsample == null) return "N/A"
-        return CORLAsample.toInt().toString()
-    }
+    // all the contests in this county
+    class CountyPoolContestBean(val countyBean: CountyPoolsBean, val info: ContestInfo, val contestTab: ContestTabulation, val isMvrs: Boolean) {
+        val contestId: Int
+        val contestName: String
+        val countyPoolId = countyBean.countyPoolId
+        val vunderTab: Vunder
+        var acBean: CountyPoolContestBean? = null
 
-    fun getCorlaHaveMvrs(): String {
-        if (contestUA == null) return "N/A"
-        val CORLAsample = contestUA!!.contest.info().metadata.get("CORLAhaveMvrs")
-        if (CORLAsample == null) return "N/A"
-        return CORLAsample.toInt().toString()
-    }
+        init {
+            contestId = contestTab.contestId
+            contestName = info.name
+            vunderTab = contestTab.votesAndUndervotes(countyPoolId, contestTab.ncards(), true)
+        }
 
-    // same as corlaSampling plot
-    fun getCorlaRisk(): Double {
-        if (contestUA == null) return 1.0
-        val cua = contestUA!!
-        val contest = cua.contest
-        val haveMvrss: String? = contest.info().metadata.get("CORLAhaveMvrs")
-        if (haveMvrss == null) return 1.0
-        return estRiskStandardBet(cua.Npop, getNoerror(), haveMvrss.toInt())
-    }
+        val undervotes = vunderTab.undervotes  // vunder properly calculates when voteForN > 1
+        val uvPct = undervotes / (contestTab.voteForN * contestTab.ncards()).toDouble()
+        val voteForN = contestTab.voteForN
+        val missing = vunderTab.missing
+        val votes = vunderTab.cands().toString()
 
-    //                 val haveMvrss: String = it.contest.info().metadata.get("CORLAhaveMvrs")!!
-    //                val haveMvrs = haveMvrss.toInt()
-    //                val estRisk = estRiskStandardBet(it.Npop, noerror, haveMvrs)
-    //                plotData.add(PlotData(it.id, noerror, estRisk, cat))
+        val nvotes = vunderTab.nvotes
+        fun getDiffNvotes() : Int {
+            return if (acBean == null) -1 else (acBean!!.nvotes - vunderTab.nvotes)
+        }
+        fun getPctDiffNvotes() : String {
+            return if (acBean == null) "" else dfn((acBean!!.nvotes - vunderTab.nvotes) / acBean!!.nvotes.toDouble(), 4)
+        }
 
-    fun getTarget(): String {
-        return if (targeted()) "YES" else ""
-    }
+        val ncards = contestTab.ncards()
+        val source = if (isMvrs) "cvrs" else "auditcenter"
 
-    fun contained(): Boolean {
-        if (contestUA == null) return false
-        val CORLAcounties = contestUA!!.contest.info().metadata.get("CORLAcounties")
-        if (CORLAcounties == null) return false
-        val toks: List<String> = CORLAcounties.split(",".toRegex()).dropLastWhile { it.isEmpty() }
-        return  (toks.size == 1)
-    }
+        fun getNCounties(): String {
+            val CORLAcounties = info.metadata.get("CORLAcounties")
+            if (CORLAcounties == null) return "N/A"
+            val toks: List<String> = CORLAcounties.split(",".toRegex()).dropLastWhile { it.isEmpty() }
+            if (toks.size == 1) return toks[0]
+            return String.format("%02d", toks.size)
+        }
 
-    fun targeted(): Boolean {
-        if (contestUA == null) return false
-        val reason = contestUA!!.contest.info().metadata.get("CORLAauditReason")
-        if (reason == null) return false
-        return reason == "state_wide_contest" || reason == "county_wide_contest"
-    }
+        fun show() = buildString {
+            append("${nfn(contestId, 3)}, ${trunc(contestName, 40)},    ${nfn(contestTab.ncards(), 6)}, ") // ${trunc(votes, 25)}, ")
+            append("    ${nfn(nvotes, 6)},   ${nfn(undervotes, 6)},    ${dfn(uvPct, 2)}")
+        }
 
-    fun isInclude(): Boolean {
-        return contestRound != null && contestRound!!.included
-    }
+        companion object {
+            val header = " id,                                     name, tabNCards, styleNCards, diffNCards, nvotes, undervotes, uvPct"
 
-    fun getMaxRisk(): String {
-        return dfn(maxrisk(), 4)
-    }
+            @JvmStatic
+            fun hiddenProperties() = "countyBean contestTab vunderTab info acBean"
 
-    fun maxrisk(): Double {
-        if (contestRound == null) return 0.0
-        val risk = contestRound!!.auditorWantRisk
-        return if (risk != null) risk else auditRiskLimit
-    }
+            @JvmStatic
+            val beanProperties = listOf(
+                TableBeanProperty("contestId", "contest Id"),
+                TableBeanProperty("contestName", "contest name"),
+                TableBeanProperty("diffNVotes", "auditcenter.votes - cvr.votes"),
+                TableBeanProperty("pctDiffNVotes", "(auditcenter.votes - cvr.votes)/auditcenter.votes"),
+            )
 
-    fun getEstRisk(): String {
-        val haveMvrs: Int = havemvrs()
-        return dfn(estRiskStandardBet(getContestPopulation(), getNoerror(), haveMvrs), 4)
-    }
-
-    fun getEstMvrs(): Int {
-        if (contestUA == null) return 0
-        return estSampleSizeStandardBet(contestUA!!.population(), getNoerror(), maxrisk())
-    }
-
-    fun getHaveMvrs(): String {
-        return nfn(havemvrs(), -6)
-    }
-
-    fun havemvrs(): Int {
-        return if (contestRound == null) 0 else contestRound!!.haveSampleSize
-    }
-
-    companion object {
-        @JvmStatic
-        fun hiddenProperties() = "countyContestData npop nmvrsConsistent nmvrsUniform contestUA contestRound auditRiskLimit";
+        }
     }
 }
